Controller implementation + tests without reflector
