+++ deployment/Cargo.toml
@@ -14,6 +14,7 @@ serde = "~1.0"
 serde_yaml = "~0.8"
 serde_json = "~1.0"
 tokio = { version = "0.2.22", features = ["macros", "rt-core"] }
+log = "0.4.11"
 
 [dev-dependencies]
 tests_common = {path = "../tests_common" }
