--- deployment/src/templates/mod.rs
@@ -1,174 +1,0 @@
-/// Contains YAML templates for Kubernetes deployment used by H2O cluster.
-/// The templates have corresponding methods to fill the template and return a new instance of
-/// respective Kubernetes component.
-use k8s_openapi::api::apps::v1::StatefulSet;
-use k8s_openapi::api::core::v1::Service;
-use k8s_openapi::api::networking::v1beta1::Ingress;
-use k8s_openapi::apiextensions_apiserver::pkg::apis::apiextensions::v1::CustomResourceDefinition;
-use serde_yaml;
-
-// TODO: Move each resource to a respective submodule
-const STATEFUL_SET_TEMPLATE: &str = r#"
-apiVersion: apps/v1
-kind: StatefulSet
-metadata:
-  name: <name>-stateful-set
-  namespace: <namespace>
-spec:
-  serviceName: h2o-service
-  podManagementPolicy: "Parallel"
-  replicas: <nodes>
-  selector:
-    matchLabels:
-      app: <name>
-  template:
-    metadata:
-      labels:
-        app: <name>
-    spec:
-      containers:
-        - name: <name>
-          image: '<docker-img-name>:<docker-img-tag>'
-          command: ["/bin/bash", "-c", "java -XX:+UseContainerSupport -XX:MaxRAMPercentage=<memory-percentage> -jar /opt/h2oai/h2o-3/h2o.jar"]
-          ports:
-            - containerPort: 54321
-              protocol: TCP
-          readinessProbe:
-            httpGet:
-              path: /kubernetes/isLeaderNode
-              port: 8081
-            initialDelaySeconds: 5
-            periodSeconds: 5
-            failureThreshold: 1
-          resources:
-            limits:
-              cpu: '<num-cpu>'
-              memory: <memory>
-            requests:
-              cpu: '<num-cpu>'
-              memory: <memory>
-          env:
-          - name: H2O_KUBERNETES_SERVICE_DNS
-            value: <name>-service.<namespace>.svc.cluster.local
-          - name: H2O_NODE_LOOKUP_TIMEOUT
-            value: '180'
-          - name: H2O_NODE_EXPECTED_COUNT
-            value: '<nodes>'
-          - name: H2O_KUBERNETES_API_PORT
-            value: '8081'
-"#;
-
-pub fn h2o_stateful_set(name: &str, namespace: &str, docker_img_name: &str, docker_img_tag: &str, nodes: u32,
-                        memory_percentage: u8, memory: &str, num_cpu: u32) -> StatefulSet {
-    let stateful_set_definition = STATEFUL_SET_TEMPLATE.replace("<name>", name)
-        .replace("<namespace>", namespace)
-        .replace("<docker-img-name>", docker_img_name)
-        .replace("<docker-img-tag>", docker_img_tag)
-        .replace("<nodes>", &nodes.to_string())
-        .replace("<memory-percentage>", &memory_percentage.to_string())
-        .replace("<memory>", memory)
-        .replace("<num-cpu>", &num_cpu.to_string());
-
-    let stateful_set: StatefulSet = serde_yaml::from_str(&stateful_set_definition).unwrap();
-    return stateful_set;
-}
-
-const SERVICE_TEMPLATE: &str = r#"
-apiVersion: v1
-kind: Service
-metadata:
-  name: <name>-service
-  namespace: <namespace>
-spec:
-  type: ClusterIP
-  clusterIP: None
-  selector:
-    app: <name>
-  ports:
-  - protocol: TCP
-    port: 80
-    targetPort: 54321
-"#;
-
-pub fn h2o_service(name: &str, namespace: &str) -> Service {
-    let service_definition = SERVICE_TEMPLATE.replace("<name>", name)
-        .replace("<namespace>", namespace);
-
-    let service: Service = serde_yaml::from_str(&service_definition).unwrap();
-    return service;
-}
-
-const INGRESS_TEMPLATE: &str = r#"
-apiVersion: networking.k8s.io/v1beta1
-kind: Ingress
-metadata:
-  name: <name>-ingress
-  annotations:
-    nginx.ingress.kubernetes.io/rewrite-target: /$2
-    traefik.frontend.rule.type: PathPrefixStrip
-spec:
-  rules:
-  - http:
-      paths:
-      - path: /<name>
-        pathType: Exact
-        backend:
-          serviceName: <name>-service
-          servicePort: 80
-"#;
-
-pub fn h2o_ingress(name: &str, namespace: &str) -> Ingress {
-    let ingress_definition = INGRESS_TEMPLATE.replace("<name>", name)
-        .replace("<namespace>", namespace);
-
-    let ingress: Ingress = serde_yaml::from_str(&ingress_definition).unwrap();
-    return ingress;
-}
-
-
-const H2O_RESOURCE_TEMPLATE: &str = r#"
-apiVersion: apiextensions.k8s.io/v1
-kind: CustomResourceDefinition
-metadata:
-  name: h2os.h2o.ai
-spec:
-  group: h2o.ai
-  names:
-    kind: H2O
-    plural: h2os
-    singular: h2o
-  scope: Namespaced
-  versions:
-    - name: v1
-      served: true
-      storage: true
-      schema:
-        openAPIV3Schema:
-          type: object
-          properties:
-            spec:
-              type: object
-              properties:
-                nodes:
-                  type: integer
-                resources:
-                  type: object
-                  properties:
-                    cpu:
-                      type: integer
-                      minimum: 1
-                    memory:
-                      type: string
-                      pattern: "^([+-]?[0-9.]+)([eEinumkKMGTP]*[-+]?[0-9]*)$"
-                    memoryPercentage:
-                      type: integer
-                      minimum: 1
-                      maximum: 100
-                  required: ["cpu", "memory"]
-              required: ["nodes", "resources"]
-"#;
-// TODO: Fill the pattern dynamically from a unified place. It is already present in CLI - make it part of Deployment ?
-
-pub fn h2o_crd() -> CustomResourceDefinition {
-    return serde_yaml::from_str(H2O_RESOURCE_TEMPLATE).unwrap();
-}
